requirements.txt
click~=8.0.0
colorthief~=0.2.1
Pillow~=8.2.0
progress~=1.5
scikit-learn~=0.24.2

# Build
pyinstaller~=4.3